Update StorageController and add tests

+++ services/storage-service/src/controllers/StorageController.js
@@ -5,7 +5,7 @@ import { logger } from '@cosy/logger';
 import CustomError from '@cosy/custom-error';
 import { tokens, refresh } from '@cosy/auth';
 import { axiosErrorBoomifier, makeAxiosInstance } from '@cosy/axios-utils';
-import StorageFactory from '../utils/StorageFactory';
+import StorageFactory from '../lib/StorageFactory';
 
 const { PODCAST_SERVICE_NAME, PODCAST_SERVICE_PORT, NODE_ENV } = process.env;
 const storages = StorageFactory();
@@ -109,23 +109,13 @@ export const getPodcastPartFile = async (id, h) => {
       return h.response(stream).type(data.contentType);
     })
     .catch((error) => {
-      if (error.isAxiosError) {
-        logger.error("Get Podcast Part File Error: Couldn't get part from podcast service", error);
-        return axiosErrorBoomifier(error);
-      }
       if (error.code === 'ENOENT' || error.statusCode === 404) {
         logger.error("Get Podcast Part File Error: Part doesn't exist", error);
-        return Boom.resourceGone('File has been deleted in storage');
+        return Boom.resourceGone("File doesn't exist or has been deleted in storage");
       }
-      // Any other error
+
       logger.error('Get Podcast Part File Error', error);
-      if (error.isBoom) {
-        return error;
-      }
-      if (error.statusCode) {
-        return new Boom.Boom(error.message, { statusCode: error.statusCode });
-      }
-      return Boom.boomify(error);
+      return axiosErrorBoomifier(error);
     });
 };
 
@@ -139,8 +129,12 @@ export const getPodcastPartFile = async (id, h) => {
  *
  * @return {Object}
  */
-export const removePodcastPartFile = async ({ storageType, storagePath, storageFilename }) => {
+export const removePodcastPartFile = async ({ storageType, storagePath, storageFilename } = {}) => {
   try {
+    if (!storageType || !storagePath || !storageFilename) {
+      throw Boom.badRequest();
+    }
+
     await storages.removeFile(storageType, storagePath, storageFilename);
     return { deleted: storageFilename };
   } catch (error) {
@@ -163,8 +157,7 @@ export const removePodcastPartFile = async ({ storageType, storagePath, storageF
  * @return {Promise<Object>}
  */
 export const addCraftFile = async ({ file, filename, storageStrategy: _storageStrategy }) => {
-  const storagePath = `crafts`;
-  const location = `${storagePath}/${filename}`;
+  const location = 'crafts';
   const defaultStorageStrategy = ['local'];
   let storageStrategy;
 
@@ -190,7 +183,11 @@ export const addCraftFile = async ({ file, filename, storageStrategy: _storageSt
   }
 
   try {
-    const storedFile = await storages.setFileFromReadable(storageStrategy, file, location);
+    const storedFile = await storages.setFileFromReadable(
+      storageStrategy,
+      file,
+      `${location}/${filename}`
+    );
 
     return {
       filename,
@@ -241,24 +238,13 @@ export const getCraftFile = async (id, h) => {
         .header('content-disposition', `attachment; filename=${_.snakeCase(data.name)}.mp3;`);
     })
     .catch((error) => {
-      if (error.isAxiosError) {
-        logger.error("Get Craft File Error: Couldn't get craft from podcast service", error);
-        return axiosErrorBoomifier(error);
-      }
       if (error.code === 'ENOENT' || error.statusCode === 404) {
         logger.error("Get Craft File Error: Craft doesn't exist", error);
+        return Boom.resourceGone("File doesn't exist or has been deleted in storage");
+      }
 
-        return Boom.resourceGone('File has been deleted in storage');
-      }
-      // Any other error
       logger.error('Get Craft File Error', error);
-      if (error.isBoom) {
-        return error;
-      }
-      if (error.statusCode) {
-        return new Boom.Boom(error.message, { statusCode: error.statusCode });
-      }
-      return Boom.boomify(error);
+      return axiosErrorBoomifier(error);
     });
 };
 
@@ -272,8 +258,12 @@ export const getCraftFile = async (id, h) => {
  *
  * @return {Object}
  */
-export const removeCraftFile = async ({ storageType, storagePath, storageFilename }) => {
+export const removeCraftFile = async ({ storageType, storagePath, storageFilename } = {}) => {
   try {
+    if (!storageType || !storagePath || !storageFilename) {
+      throw Boom.badRequest();
+    }
+
     await storages.removeFile(storageType, storagePath, storageFilename);
     return { deleted: storageFilename };
   } catch (error) {
